requirements.txt
discord.py>=2.5,<3.0
python-dotenv>=1.0,<2.0
redis[hiredis]>=6.0,<7.0
